# packages/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Fakegaming Bot API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API base path"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Login via /auth/login and paste the token here using the \"Authorize\" button."
      }
    },
    "schemas": {
...
      "PatchNoteConfig": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "game": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "publishedAt": {
            "type": "integer",
            "format": "int64"
          },
          "logoUrl": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "accentColor": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "PatchSubscriptionConfig": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "game": {
            "type": "string"
          },
          "channelId": {
            "type": "string"
          },
          "guildId": {
            "type": "string"
          },
          "lastAnnouncedAt": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "QuoteConfig": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "guildId": {
            "type": "string"
          },
          "quote": {
            "type": "string"
          },
          "authorId": {
            "type": "string"
          },
          "submitterId": {
            "type": "string"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/patchNotes": {
      "get": {
        "summary": "List all patch notes",
        "tags": [
          "PatchNotes"
        ],
        "responses": {
          "200": {
            "description": "List of patch notes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatchNoteConfig"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Upsert (add or update) a patch note",
        "tags": [
          "PatchNotes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchNoteConfig"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/patchNotes/{game}": {
      "get": {
        "summary": "Get the latest patch note for a game",
        "tags": [
          "PatchNotes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "game",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patch note config",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchNoteConfig"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/patchSubscriptions": {
      "get": {
        "summary": "List all patch subscriptions",
        "tags": [
          "PatchSubscriptions"
        ],
        "responses": {
          "200": {
            "description": "List of patch subscriptions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatchSubscriptionConfig"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Subscribe to a game/channel",
        "tags": [
          "PatchSubscriptions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchSubscriptionConfig"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "put": {
        "summary": "Upsert (add or update) a patch subscription",
        "tags": [
          "PatchSubscriptions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchSubscriptionConfig"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/quotes": {
      "get": {
        "summary": "List all quotes",
        "tags": [
          "Quotes"
        ],
        "responses": {
          "200": {
            "description": "List of quotes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuoteConfig"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new quote",
        "tags": [
          "Quotes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteConfig"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/quotes/search": {
      "get": {
        "summary": "Search quotes by text and guildId",
        "tags": [
          "Quotes"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "guildId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "text",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of quotes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuoteConfig"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/quotes/guild/{guildId}": {
      "get": {
        "summary": "Get quotes by guild",
        "tags": [
          "Quotes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "guildId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of quotes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuoteConfig"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/quotes/guild/{guildId}/author/{authorId}": {
      "get": {
        "summary": "Get quotes by author in guild",
        "tags": [
          "Quotes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "guildId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "authorId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of quotes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuoteConfig"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/quotes/{id}": {
      "delete": {
        "summary": "Delete a quote by id",
        "tags": [
          "Quotes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "get": {
        "summary": "Get a quote by id",
        "tags": [
          "Quotes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A quote object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteConfig"
                }
              }
            }
          },
          "404": {
            "description": "Quote not found"
          }
        }
      }
    },
...
  },
...
}